Use prefix form of :func: and similar tags. (#17)
Although the suffix form of role markers such as :func: and :class: are allowed by the reStructuredText specification ("A role marker is either a prefix or a suffix to the interpreted text, whichever reads better; it's up to the author"), some tools don't support it.  This revision converts all suffix references to prefix references.